--- src/common/components/Layout/components/Header/Mobile/index.tsx
@@ -1,5 +1,0 @@
-const MobileHeader = () => {
-  return <header>index</header>;
-};
-
-export default MobileHeader;
