+++ src/common/components/Layout/components/Header/Nav/style.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export const menuList = style({
+  display: 'flex',
+  alignItems: 'center',
+  gap: 40,
+});
